refactor(mcps): address ajaen-ai's July 16 review nits
- output tests (mcps_test.go) now compare exact expected strings
  instead of strings.Contains, matching the rest of the repo
- TestBuildMcpRequestBody flattens wantErr bool + errContain into a
  single wantErr string; TestBuildMcpUpdatePatch rewritten as a proper
  table-driven test with an exact want map compared via cmp.Diff
- sendMcpRequest promoted to a generic DeploymentClient.sendJSONRequest
  (lives in deployment_client.go, no "Mcp" in the name); DescribeMcp now
  calls it instead of duplicating the request/response handling
- mcp_create.go merged into mcp.go (was two small files for one
  cohesive concern); test files merged the same way
